Use oi instance instead of static class
The robot creates a new OI instance that we put the button actions in,
so we do have a place to store the other things such as the controller.
Moved the controller to private -- other objects should be asking OI for
the information they need and not hitting the hardware directly.